# src/main/java/com/example/bankcards/dto/UserDTO.java
package com.example.bankcards.dto;

import com.example.bankcards.entity.User;

public class UserDTO {
    private Integer id;
    private String username;
    private String role;

    public Integer getId() {
        return id;
    }

    public void setId(Integer id) {
        this.id = id;
    }

    public String getRole() {
        return role;
    }

    public void setRole(String role) {
        this.role = role;
    }

    public String getUsername()  {
        return username;
    }

    public void setUsername(String username) {
        this.username = username;
    }

    public UserDTO() {};

    public UserDTO(User user) {
        this.id = user.getId();
        this.username = user.getUsername();
        this.role = user.getRole();
    }
}